requirements.txt
pillow
google-generativeai
playwright
beautifulsoup4
requests
PyMuPDF
markdownify
reportlab
backoff
gradio